JWT requirements and admin priveleges for post/put/delete endpoints

+++ resources/game.py
@@ -1,4 +1,5 @@
 from flask.views import MethodView
+from flask_jwt_extended import get_jwt, jwt_required
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -16,12 +17,16 @@ class Game(MethodView):
     def get(self, game_id):
         return GameModel.query.get_or_404(game_id)
     
+    # TODO: Put in some kind of restriction/rule of who can modify an existing game's info. Maybe just admins once game created?
+    @jwt_required()
     @blp.arguments(GameUpdateSchema)
     @blp.response(201, GameSchema)
     def put(self, request_payload, game_id):
         game = GameModel.query.get(game_id)
         if game:
-            game.game_name = request_payload['game_name']
+            game.game_name = request_payload['game_name'] if 'game_name' in request_payload else game.game_name
+            game.wiki_url = request_payload['wiki_url'] if 'wiki_url' in request_payload else game.wiki_url
+            game.image_url = request_payload['image_url'] if 'image_url' in request_payload else game.image_url
         else:
             game = GameModel(**request_payload)
         try:
@@ -31,13 +36,20 @@ class Game(MethodView):
         except SQLAlchemyError:
             abort(500, message="Unable to post to SQL database.")
 
-    @blp.response(200, GameSchema)
+    # TODO: Rather than actually delete, just flag it for deletion later so it's not a true delete
+    # TODO: Flag game/word links to be deleted as well.
+    @jwt_required()
+    @blp.response(204)
     def delete(self, game_id: str):
+        jwt = get_jwt()
+        if not jwt.get('is_admin'):
+            abort(403, message="Permission denied. Admin privelege required.")
+
+        game = GameModel.query.get_or_404(game_id)
         try:
-            game = GameModel.query.get_or_404(game_id)
             db.session.delete(game)
             db.session.commit()
-            return game
+            return {}
         except KeyError:
             abort(404, message=f"Game with ID {game_id} not found.")
 
@@ -48,6 +60,7 @@ class GameList(MethodView):
     def get(self):
         return GameModel.query.all()
 
+    @jwt_required()
     @blp.arguments(GameSchema)
     @blp.response(201, GameSchema)
     def post(self, request_payload):
@@ -73,6 +86,8 @@ class GameWordsList(MethodView):
 
 @blp.route("/game/<int:game_id>/word/<int:word_id>")
 class LinkGameToWord(MethodView):
+    # TODO: Decide if this is allowable by anyone or any user with token.
+    @jwt_required()
     @blp.response(201, WordSchema)
     def post(self, game_id: int, word_id: int):
         game = GameModel.query.get_or_404(game_id)
@@ -86,8 +101,15 @@ class LinkGameToWord(MethodView):
         except SQLAlchemyError:
             abort(500, message="Unable to link game and word in database.")
     
-    @blp.response(200, WordSchema)
+    # TODO: Figure out some kind of rule/permission restriction so not anyone can make these deletes.
+    @jwt_required()
+    @blp.response(204)
     def delete(self, game_id: int, word_id: int):
+
+        jwt = get_jwt()
+        if not jwt.get('is_admin'):
+            abort(401, message="Permission denied. Admin privelege required.")
+
         game = GameModel.query.get_or_404(game_id)
         word = WordModel.query.get_or_404(word_id)
 
@@ -95,6 +117,6 @@ class LinkGameToWord(MethodView):
         try:
             db.session.add(game)
             db.session.commit()
-            return word
+            return {}
         except SQLAlchemyError:
             abort(500, message="Unable to delete game and word link in database.")
